- Cleaned up the interfaces for several of the resources, in particular Image, Font and Sound. - Modified Font so that styles (like bold, italic and underline) are toggled via function calls. Removes the FontStyle enumerator. - Cleaned a warning in Image by explicitly type casting an unsigned int to an int. - Pulled 'texture id' code out of Image... it doesn't belong there at all. - Fixed a potential bug in SDL_Renderer where the surface flags for the video screen was never initialized. - Removed reference to Image's deprecated textureId() function from OGL_Renderer.

+++ core/src/Resources/Sprite.cpp
@@ -157,7 +157,7 @@ void Sprite::update(int x, int y)
 	if(!mPaused && (frame.frameDelay() != FRAME_PAUSE))
 	{
 		unsigned int currentTick = mTimer.ms();
-		if(currentTick - mLastFrameTick > frame.frameDelay())
+		if(static_cast<int>(currentTick - mLastFrameTick) > frame.frameDelay())
 		{
 			mLastFrameTick = currentTick;
 			mCurrentFrame++;
